Obtido conteúdo de um livro

+++ src/main/java/br/com/cabidiomas/library/book/repository/BookRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface BookRepository  extends JpaRepository<Book, Long> {
 
-    List<Book> findAllByLevelId(Long id);
+    List<Book> findAllByLevelId(Integer id);
 }
